requirements.txt
# ==============================================================================
# 氢能专利知识图谱研究平台 - 依赖列表
# ==============================================================================

# 环境变量支持
python-dotenv>=1.0.0

# ==============================================================================
# 核心依赖
# ==============================================================================
pandas>=2.0.0
numpy>=1.23.5
openpyxl>=3.0.10
requests>=2.28.0
tqdm>=4.65.0
pyyaml>=6.0

# ==============================================================================
# Neo4j 图数据库
# ==============================================================================
neo4j>=5.0.0
neo4j-graphrag>=0.3.0

# ==============================================================================
# 向量搜索
# ==============================================================================
faiss-cpu>=1.7.0
sentence-transformers>=2.2.0

# ==============================================================================
# LangGraph Agent（知识图谱问答）
# ==============================================================================
langgraph>=0.2.0
langchain>=0.2.0
langchain-core>=0.2.0
langchain-community>=0.2.0

# ==============================================================================
# 深度学习框架（模型评测/本地推理）
# ==============================================================================
torch>=2.0.0
transformers>=4.36.0

# 本地vLLM推理（可选，如需本地模型评测）
# vllm>=0.3.0

# ==============================================================================
# 可视化（可选）
# ==============================================================================
matplotlib>=3.7.0
